requirements.txt
Pillow==6.1.0
django==2.2.5
djangorestframework==3.10.3
django-cors-headers==3.1.1